Ruby: Move common stuff to shared module.
