+++ ruby/optify/lib/optify_ruby/watcher_implementation.rb
@@ -8,23 +8,11 @@ require 'sorbet-runtime'
 require_relative './base_config'
 require_relative './implementation'
 require_relative './options_metadata'
+require_relative './provider_module'
 
 module Optify
-  # Like `OptionsProvider` but also watches for changes to the files and reloads the options.
-  class OptionsWatcher
-    extend T::Sig
-
-    #: (Array[String] feature_names) -> Array[String]
-    def get_canonical_feature_names(feature_names)
-      # Try to optimize a typical case where there are just a few features.
-      # Ideally in production, a single feature that imports many other features is used for the most common scenario.
-      # Benchmarks show that it is faster to use a loop than to call the Rust implementation which involves making a `Vec<String>` and returning a `Vec<String>`.
-      if feature_names.length < 4
-        feature_names.map { |feature_name| get_canonical_feature_name(feature_name) }
-      else
-        _get_canonical_feature_names(feature_names)
-      end
-    end
+  class OptionsWatcher # rubocop:disable Style/Documentation
+    include ProviderModule
 
     #: -> Hash[String, OptionsMetadata]
     def features_with_metadata
@@ -33,14 +21,6 @@ module Optify
         result[key] = OptionsMetadata.from_hash(value)
       end
       result
-    end
-
-    #: (String canonical_feature_name) -> Optify::OptionsMetadata?
-    def get_feature_metadata(canonical_feature_name)
-      metadata_json = get_feature_metadata_json(canonical_feature_name)
-      return nil if metadata_json.nil?
-
-      OptionsMetadata.from_hash(JSON.parse(metadata_json))
     end
 
     # Fetches options based on the provided key and feature names.
