+++ ruby/optify/sig/optify.rbs
@@ -40,8 +40,34 @@ class Optify::GetOptionsPreferences
   def skip_feature_name_conversion: () -> bool
 end
 
+# A module that provides the methods to help implement providers.
+module Optify::ProviderModule
+  # Map an alias or canonical feature name (perhaps derived from a file name) to a canonical feature name.
+  # Canonical feature names map to themselves.
+  #
+  # @param feature_name The name of an alias or a feature.
+  # @return The canonical feature name.
+  def get_canonical_feature_name: (String feature_name) -> String
+
+  # Map aliases or canonical feature names (perhaps derived from a file names) to the canonical feature names.
+  # Canonical feature names map to themselves.
+  # This implementation calls the Rust implementation directly.
+  #
+  # @param feature_names The names of aliases or features.
+  # @return The canonical feature names.
+  def _get_canonical_feature_names: (::Array[String] feature_names) -> ::Array[String]
+
+  # @return The metadata for the feature.
+  def get_feature_metadata_json: (String canonical_feature_name) -> String?
+
+  # @return All of the keys and values for the the features.
+  def features_with_metadata_json: () -> String
+end
+
 # Provides configurations based on keys and enabled feature names.
 class Optify::OptionsProvider
+  include ProviderModule
+
   # @return All of the canonical feature names.
   def features: () -> ::Array[String]
 
@@ -49,13 +75,6 @@ class Optify::OptionsProvider
   def features_with_metadata: () -> ::Hash[String, OptionsMetadata]
 
   def get_all_options_json: (::Array[String] feature_names, GetOptionsPreferences preferences) -> String
-
-  # Map an alias or canonical feature name (perhaps derived from a file name) to a canonical feature name.
-  # Canonical feature names map to themselves.
-  #
-  # @param feature_name The name of an alias or a feature.
-  # @return The canonical feature name.
-  def get_canonical_feature_name: (String feature_name) -> String
 
   # Map aliases or canonical feature names (perhaps derived from a file names) to the canonical feature names.
   # Canonical feature names map to themselves.
@@ -82,20 +101,6 @@ class Optify::OptionsProvider
   # (Optional) Eagerly initializes the cache.
   # @return [OptionsProvider] `self`.
   def init: () -> OptionsProvider
-
-  # Map aliases or canonical feature names (perhaps derived from a file names) to the canonical feature names.
-  # Canonical feature names map to themselves.
-  # This implementation calls the Rust implementation directly.
-  #
-  # @param feature_names The names of aliases or features.
-  # @return The canonical feature names.
-  def _get_canonical_feature_names: (::Array[String] feature_names) -> ::Array[String]
-
-  # @return The metadata for the feature.
-  def get_feature_metadata_json: (String canonical_feature_name) -> String?
-
-  # @return All of the keys and values for the the features.
-  def features_with_metadata_json: () -> String
 end
 
 # A builder for creating an `OptionsProvider` instance.
@@ -119,13 +124,6 @@ class Optify::OptionsWatcher
   def features_with_metadata: () -> ::Hash[String, OptionsMetadata]
 
   def get_all_options_json: (::Array[String] feature_names, GetOptionsPreferences preferences) -> String
-
-  # Map an alias or canonical feature name (perhaps derived from a file name) to a canonical feature name.
-  # Canonical feature names map to themselves.
-  #
-  # @param feature_name The name of an alias or a feature.
-  # @return The canonical feature name.
-  def get_canonical_feature_name: (String feature_name) -> String
 
   # Map aliases or canonical feature names (perhaps derived from a file names) to the canonical feature names.
   # Canonical feature names map to themselves.
@@ -152,20 +150,6 @@ class Optify::OptionsWatcher
   # (Optional) Eagerly initializes the cache.
   # @return [OptionsWatcher] `self`.
   def init: () -> OptionsWatcher
-
-  # Map aliases or canonical feature names (perhaps derived from a file names) to the canonical feature names.
-  # Canonical feature names map to themselves.
-  # This implementation calls the Rust implementation directly.
-  #
-  # @param feature_names The names of aliases or features.
-  # @return The canonical feature names.
-  def _get_canonical_feature_names: (::Array[String] feature_names) -> ::Array[String]
-
-  # @return The metadata for the feature.
-  def get_feature_metadata_json: (String canonical_feature_name) -> String?
-
-  # @return All of the keys and values for the the features.
-  def features_with_metadata_json: () -> String
 end
 
 # A builder for creating an `OptionsWatcher` instance.
